delete / CardWithLabel Dao and move function
