+++ data/db/src/main/java/com/techcafe/todone/db/internal/dao/CardEntityDao.kt
@@ -2,6 +2,8 @@ package com.techcafe.todone.db.internal.dao
 
 import androidx.room.*
 import com.techcafe.todone.db.internal.entity.CardEntity
+import com.techcafe.todone.db.internal.entity.LabelEntity
+import com.techcafe.todone.db.internal.middleEntity.CardWithLabel
 
 /**
  * カードに関するDao
@@ -16,6 +18,15 @@ interface CardEntityDao {
      */
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     fun insertCard(vararg card: CardEntity)
+
+    /**
+     * カードとラベルを紐づける関数
+     *
+     * @param [CardWithLabel] 登録したいカードとラベルのidを持つCardWithLabelインスタンス
+     * @see CardWithLabel
+     */
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun bindLabel(cardWithLabel: CardWithLabel)
 
     /**
      * 看板に紐づいているカードのリストを返す関数
@@ -36,4 +47,21 @@ interface CardEntityDao {
     @Transaction
     @Query("SELECT * FROM card_item WHERE card_id = :cardId LIMIT 1")
     suspend fun getCardById(cardId: Int): CardEntity?
+
+    /**
+     * カードに紐づいているラベルを取得する関数
+     *
+     * @param [cardId] 紐付くラベル一覧を取得したいカードのid
+     * @see LabelEntity
+     * @see CardWithLabel
+     */
+    @Query(
+        """
+               SELECT * FROM label_item
+               INNER JOIN card_with_label
+               ON label_item.label_id=card_with_label.label_id_with
+               WHERE card_with_label.card_id_with=:cardId
+               """
+    )
+    fun getLabelsForCard(cardId: Int): List<LabelEntity>
 }
