+++ data/db/src/main/java/com/techcafe/todone/db/internal/dao/LabelEntityDao.kt
@@ -4,7 +4,9 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.techcafe.todone.db.internal.entity.CardEntity
 import com.techcafe.todone.db.internal.entity.LabelEntity
+import com.techcafe.todone.db.internal.middleEntity.CardWithLabel
 
 /**
  * ラベルに関するDao
@@ -27,4 +29,21 @@ interface LabelEntityDao {
      */
     @Query("SELECT * FROM label_item")
     fun getLabelList(): List<LabelEntity>
+    
+    /**
+     * ラベルに紐付くカードのリストを取得する関数
+     *
+     * @param [labelId] 紐付くカード一覧を取得したいラベルのid
+     * @see CardEntity
+     * @see CardWithLabel
+     */
+    @Query(
+        """
+               SELECT * FROM card_item
+               INNER JOIN card_with_label
+               ON card_item.card_id=card_with_label.card_id_with
+               WHERE card_with_label.label_id_with=:labelId
+               """
+    )
+    fun getCardsForLabel(labelId: Int): List<CardEntity>
 }
